non shared changes and disabling tests
