+++ tests/TopN.CoreFX.Windows.issues.json
@@ -1930,6 +1930,14 @@
                     "reason": "Xunit.Sdk.EqualException"
                 },
                 {
+                    "name": "System.Threading.Tests.EventWaitHandleTests.OpenExisting_InvalidNames_Windows",
+                    "reason": "Xunit.Sdk.EqualException"
+                },
+                {
+                    "name": "System.Threading.Tests.EventWaitHandleTests.Ctor_InvalidNames",
+                    "reason": "Xunit.Sdk.EqualException"
+                },
+                {
                     "name": "System.Threading.Tests.MonitorTests.PulseAll_Invalid",
                     "reason": "Xunit.Sdk.EqualException"
                 },
@@ -1948,6 +1956,22 @@
                 {
                     "name": "System.Threading.Tests.SemaphoreTests.PingPong",
                     "reason": "System.TypeInitializationException"
+                },
+                {
+                    "name": "System.Threading.Tests.MutexTests.Ctor_InvalidName",
+                    "reason": "Xunit.Sdk.EqualException"
+                },
+                {
+                    "name": "System.Threading.Tests.MutexTests.OpenExisting_InvalidNames",
+                    "reason": "Xunit.Sdk.EqualException"
+                },
+                {
+                    "name": "System.Threading.Tests.SemaphoreTests.OpenExisting_InvalidNames_Windows",
+                    "reason": "Xunit.Sdk.EqualException"
+                },
+                {
+                    "name": "System.Threading.Tests.SemaphoreTests.Ctor_InvalidNames",
+                    "reason": "Xunit.Sdk.EqualException"
                 }
             ]
         }
